requirements.txt
flask
flask_sqlalchemy
flask_migrate
requests
openai
python-dotenv
